整理 objects 中的节点；整理 parse 中的方法；优化引用路径
