+++ demo_2.py
@@ -2,8 +2,7 @@
 解析 MySQL 建表语句
 """
 
-from metasequoia_sql.parser.expression import parse_create_table_statement
-from metasequoia_sql.common import build_token_scanner
+from metasequoia_sql import *
 
 if __name__ == "__main__":
     statement = parse_create_table_statement(build_token_scanner("""
