+++ metasequoia_sql/core/__init__.py
@@ -1,1 +1,3 @@
 from metasequoia_sql.core.data_source import DataSource
+from metasequoia_sql.core.objects import *
+from metasequoia_sql.core.parser import *
