--- metasequoia_sql/static/__init__.py
@@ -1,3 +1,0 @@
-from metasequoia_sql.static.common import *
-from metasequoia_sql.static.mysql import MYSQL_KEYWORDS
-from metasequoia_sql.static.other import MYSQL_DATA_TYPE, HASHMAP_MYSQL_TO_HIVE
