--- metasequoia_sql/static/common.py
@@ -1,6 +1,0 @@
-# 十六机制的字符集
-HEXADECIMAL_CHARACTER_SET = {"0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "A", "a", "B", "b", "C", "c", "D", "d",
-                             "E", "e", "F", "f"}
-
-# 二级制的字符集
-BINARY_CHARACTER_SET = {"0", "1"}
