+++ scripts/temp_test.py
@@ -1,5 +1,4 @@
-from metasequoia_sql.common import build_token_scanner
-from metasequoia_sql.parser.expression import parse_select_statement
+from metasequoia_sql import *
 from scripts.demo.sql_basic_tutorial import *
 
 statement = parse_select_statement(build_token_scanner(SBT_CH05_10_SQLSERVER))
