+++ scripts/unittest_maker.py
@@ -6,10 +6,7 @@ import json
 import os
 import subprocess
 
-from metasequoia_sql.common import build_token_scanner
-from metasequoia_sql.parser.expression import parse_select_statement, maybe_select_statement
-from metasequoia_sql.core.objects import SQLSingleSelectStatement
-from metasequoia_sql.objects import DataSource
+from metasequoia_sql import *
 from scripts.demo import sql_basic_tutorial
 
 
@@ -31,8 +28,7 @@ def make_sql_basic_tutorial(force: bool = False):
         # 生成引用信息
         file.write("import unittest\n")
         file.write("\n")
-        file.write("from metasequoia_sql.common import build_token_scanner\n")
-        file.write("from metasequoia_sql.parser.common import parse_select_statement\n")
+        file.write("from metasequoia_sql import *\n")
         file.write("from scripts.demo.sql_basic_tutorial import *\n")
         file.write("\n")
         file.write("\n")
@@ -49,7 +45,7 @@ def make_sql_basic_tutorial(force: bool = False):
             sql = getattr(sql_basic_tutorial, name)
             scanner = build_token_scanner(sql)
 
-            if not maybe_select_statement(scanner):
+            if not check_select_statement(scanner):
                 continue  # 当前只处理 SELECT 语句
 
             # 语法分析并打印结果
